draw multiple tours in one pathlayer

+++ bundles/net.tourbook/src/net/tourbook/map25/OkHttpEngineMT.java
@@ -73,7 +73,7 @@ public class OkHttpEngineMT extends OkHttpEngine {
 		 */
 		public OkHttpFactoryMT() {
 
-			_httpCache = new Cache(new File(getCacheDir()), Integer.MAX_VALUE);
+			_httpCache = new Cache(new File(getCacheDir()), Long.MAX_VALUE);
 
 			final OkHttpClient.Builder httpBuilder = new OkHttpClient.Builder();
